pyproject.toml
[tool.poetry]
name = "scheduler_bot"
version = "0.1.0"
description = ""
authors = ["Sergey Rubnikovich <sergeyrubnikovich@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
uvloop = "^0.17.0"
orjson = "^3.9.0"
aiohttp = {extras = ["speedups"], version = "^3.8.4"}
python-dotenv = "^1.0.0"
emoji = "^2.4.0"
aiogram = "3.0.0b7"

[tool.poetry.dev-dependencies]
pre-commit = "^3.3.2"
flake8 = "^6.0.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.isort]
known_local_folder = "scheduler"
lines_after_imports = 2
multi_line_output = 3
line_length = 100
include_trailing_comma = true
